Day1/pyproject.toml
[project]
name = "day1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0",
    "django-filter>=25.2",
    "django-silk>=5.4.3",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt[crypto]>=5.5.1",
    "drf-spectacular>=0.29.0",
    "pillow>=12.0.0",
]
